requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-WTF==1.2.1
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
PyPDF2==3.0.1
pytest==8.1.1
tomli==2.0.1
Werkzeug==3.0.1
WTForms==3.1.2
